Correction confirmationReservation, amélioration du panel, correction factoryClient

+++ src/data/Reservation.java
@@ -104,7 +104,7 @@ public class Reservation {
 		return prix;
 	}
 
-	public void setPrix(int prix) {
+	public void setPrix(float prix) {
 		this.prix = prix;
 	}
